app/requirements.txt
Django
djangorestframework
pylint-django
dacite
